+++ CHANGELOG.md
@@ -17,7 +17,10 @@ CHANGELOG
   [#4059](https://github.com/pulumi/pulumi/pull/4059)
 
 - Add support for stack transformations in the .NET SDK.
-  [4008](https://github.com/pulumi/pulumi/pull/4008)
+  [#4008](https://github.com/pulumi/pulumi/pull/4008)
+
+- Fix `pulumi stack ls` on Windows
+  [#4094](https://github.com/pulumi/pulumi/pull/4094)
   
 ## 1.12.1 (2020-03-11)
 - Fix Kubernetes YAML parsing error in .NET.